refactor: model-vue fixed

+++ model/src/main/java/com/apighost/model/loadtest/parameter/Stage.java
@@ -15,7 +15,7 @@ import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 public class Stage {
 
     @JsonProperty("vus")
-    private final long vusMs;
+    private final int vusMs;
     @JsonProperty("duration")
     private final long durationMs;
 
@@ -30,11 +30,11 @@ public class Stage {
     @JsonPOJOBuilder(withPrefix = "")
     public static class Builder {
 
-        private long vusMs;
+        private int vusMs;
         private long durationMs;
 
         @JsonProperty("vus")
-        public Builder vusMs(long vusMs) {
+        public Builder vusMs(int vusMs) {
             this.vusMs = vusMs;
             return this;
         }
@@ -54,7 +54,7 @@ public class Stage {
      * Getter
      */
     @JsonProperty("vus")
-    public long getVusMs() {
+    public int getVusMs() {
         return vusMs;
     }
